Implment polyfill of scheduler.yield() (#4)
This adds support for scheduler.yield() for both browsers with
scheduler (Chrome) and without.

If postTask is available, yield() is polyfilled with postTask(). This
allows us to schedule user-visible continuations at a higher event loop
priority and get similar scheduler characteristics as yield(). We can't,
however, guarantee that yield() continuations always run before
postTask() tasks of the same priority using this approach. In
particular:
 - "user-blocking" tasks and continuations, and "user-visible"
   continuations can all be interleaved.
 - "background" tasks and continuations can be interleaved.

If postTask() is not available (polyfilling both), then continuations
are guaranteed to run before postTask() tasks of the same priority, but
continuations never have a higher event loop priority.

Implmentation notes:
 - For the version integrating with the existing polyfill, since event
   loop scheduling for continuations and tasks is the same, we only need
   to make sure to run continuations before tasks. This is handled by
   storing continuations in separate task queues (alongside the existing
   ones) and choosing continuations over tasks. postTask() was
   refactored into a common function for both postTask() and yield(),
   with the queue being chosen based on the API.

 - For the postTask version, the main challenge is getting priority
   change to work, e.g. when changing priority to "user-visible", the
   underlying tasks should be scheduled as "user-blocking". For this, we
   listen for priority/abort events on the signal provided in the
   options and forward abort and priority, with the latter being
   adjusted as needed.

Additionally, this updates the README with notes about event loop
scheduling for postTask() and yield().